Revert "refactor to remove the recursion feature flag"
This reverts commit 14fb44d847c432302566aef44d18bf5b1e1fef51.
