+++ risc0/zkvm/Cargo.toml
@@ -168,6 +168,7 @@ prove = [
   "dep:sha2",
   "dep:tempfile",
   "dep:typetag",
+  "risc0-binfmt/recursion",
   "risc0-circuit-recursion/prove",
   "risc0-circuit-rv32im/prove",
   "risc0-zkp/prove",
